% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bak.vir}
\alias{bak.vir}
\title{bak.vir data set, a class of matrix.}
\format{
a class of matrix.
\describe{
\item{bak.vir}{5000*5 matrix.}
}
}
\usage{
bak.vir
}
\description{
A dataset containing 5 of the 19 bioclimatic variables randomly genereated as
background points.
}
\keyword{datasets}
