% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en.vir}
\alias{en.vir}
\title{en.vir data set, a class of RasterBrick.}
\format{
a class of RasterBrick.
\describe{
\item{en.vir}{class: RasterBrick;
dimensions : 360, 720, 259200, 5  (nrow, ncol, ncell, nlayers);
resolution : 0.5, 0.5  (x, y);
extent: -180, 180, -90, 90  (xmin, xmax, ymin, ymax);
crs: +proj=longlat +datum=WGS84 +no_defs;
source: memory;
names: bio1,bio4,bio7,bio12,bio15.
}
}
}
\source{
\url{https://www.worldclim.org/}
}
\usage{
en.vir
}
\description{
A dataset containing 5 of the 19 bioclimatic variables downloaded from WorldClim
(version 1.4 with 10 arc minute resolution; Hijmans et al. 2005)).
}
\keyword{datasets}
