% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LappetMoths}
\alias{LappetMoths}
\title{LappetMoths data set, a list of 8 data frames.}
\format{
list of 8 data frames.

\describe{
\item{barcode.identi.result}{data frame,species identifications by other
methods or barocodes,containing query IDs, species identified,
and corresponding probablities.}
\item{que.env}{data frame, containing query sampleIDs,and a set of
corresponding environmental variables collected by users.}
\item{que.infor}{data frame, query samples,containing sample IDs,longitude
and latitude of each sample.}
\item{que.seq}{query sequences in binary format stored in a matrix}
\item{ref.env}{data frame, containing reference sampleIDs, species names,
and a set of environmental variables collected by users.}
\item{ref.infor}{data frame, reference dataset containing sample IDs, taxon
information,longitude and latitude of each sample.}
\item{ref.seq}{reference sequences in binary format stored in a matrix}
\item{ref.add}{data frame, additional reference dataset containing taxon
information, longitude and latitude of each species.}
}
}
\usage{
LappetMoths
}
\description{
A dataset containing the sequences IDs of species,
coordinates of species sampled, and other attributes
}
\keyword{datasets}
