% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpeInfo.R
\name{extractSpeInfo}
\alias{extractSpeInfo}
\title{Extraction of taxon/species  and distribution information}
\usage{
extractSpeInfo(seqID.full)
}
\arguments{
\item{seqID.full}{Character, sample ID, taxon information and longitude and
latitude data that splitted by comma in class character.}
}
\value{
A data frame of splitted sample ID, taxon information and longitude
and latitude data for further analysis.
}
\description{
Split comma-separated sample information into different columns
of a data frame.
}
\examples{
data(LappetMoths)
ref.seq<-LappetMoths$ref.seq
seqID.full<-rownames(ref.seq)

infor<-extractSpeInfo(seqID.full)
head(infor)
}
\author{
Cai-qing YANG (Email: yangcq_ivy(at)163.com) and Ai-bing ZHANG
(Email:zhangab2008(at)cnu.edu.cn), Capital Normal University (CNU), Beijing,
CHINA.
}
\keyword{extractSpeInfo}
